backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Demo"
authors = ["Andrew Ivins <andrew@ivins.id.au>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.6"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
graphene-django = "^2.15.0"
django-extensions = "^3.1.3"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"
ipdb = "^0.13.9"
ipython = "^7.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
